# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DCA Strategy Assistant",
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/assets/{symbol}/range": {
      "get": {
        "summary": "Asset Range",
        "description": "Return the date range the UI can use as min/max on date inputs.\n\nSourced from the local SQLite cache when populated, falling back\nto a hardcoded earliest-available date per symbol when the cache\nis empty (fresh install). Avoids hitting yfinance just to discover\nthe floor.",
        "operationId": "asset_range_api_assets__symbol__range_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetRange"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/recommendations/run": {
      "post": {
        "summary": "Recommendation",
        "operationId": "recommendation_api_recommendations_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/explanations/run": {
      "post": {
        "summary": "Explanation",
        "description": "Generate a plain-language explanation of the current\nrecommendation via the user's OpenAI-compatible LLM.\n\nThe API key in request.llm is forwarded to the provider for this\nsingle call only — never persisted, never logged. See\nexplanations.py for the request construction.",
        "operationId": "explanation_api_explanations_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplanationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplanationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/explanations/selection": {
      "post": {
        "summary": "Selection Explanation",
        "description": "Explain user-selected page text with current strategy context.\n\nThe selected text is treated as untrusted quoted content by the\nprompt; it is never executed as an instruction. The API key follows\nthe same per-request forwarding rules as /api/explanations/run.",
        "operationId": "selection_explanation_api_explanations_selection_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectionExplanationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectionExplanationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/optimizations/run": {
      "post": {
        "summary": "Optimization",
        "operationId": "optimization_api_optimizations_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimizationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/optimizations/jobs": {
      "post": {
        "summary": "Create Optimization",
        "operationId": "create_optimization_api_optimizations_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimizationJobCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/optimizations/jobs/{job_id}": {
      "get": {
        "summary": "Optimization Status",
        "operationId": "optimization_status_api_optimizations_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimizationJobStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Cancel Optimization",
        "operationId": "cancel_optimization_api_optimizations_jobs__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimizationJobStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/backtests/run": {
      "post": {
        "summary": "Backtest",
        "operationId": "backtest_api_backtests_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AssetRange": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "minDate": {
            "type": "string",
            "format": "date",
            "title": "Mindate"
          },
          "maxDate": {
            "type": "string",
            "format": "date",
            "title": "Maxdate"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "minDate",
          "maxDate"
        ],
        "title": "AssetRange"
      },
...
      "ExplanationRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol",
            "default": "QQQ"
          },
          "config": {
            "$ref": "#/components/schemas/StrategyConfig"
          },
          "asOf": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asof"
          },
          "llm": {
            "$ref": "#/components/schemas/LlmSettings"
          }
        },
        "type": "object",
        "required": [
          "llm"
        ],
        "title": "ExplanationRequest"
      },
      "ExplanationResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "decision": {
            "$ref": "#/components/schemas/StrategyDecision"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "dataSource": {
            "type": "string",
            "title": "Datasource"
          },
          "cacheStatus": {
            "type": "string",
            "title": "Cachestatus"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "decision",
          "explanation",
          "model",
          "dataSource",
          "cacheStatus"
        ],
        "title": "ExplanationResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RecommendationRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol",
            "default": "QQQ"
          },
          "config": {
            "$ref": "#/components/schemas/StrategyConfig"
          },
          "asOf": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asof"
          }
        },
        "type": "object",
        "title": "RecommendationRequest"
      },
      "RecommendationResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "decision": {
            "$ref": "#/components/schemas/StrategyDecision"
          },
          "dataSource": {
            "type": "string",
            "title": "Datasource"
          },
          "cacheStatus": {
            "type": "string",
            "title": "Cachestatus"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "decision",
          "dataSource",
          "cacheStatus"
        ],
        "title": "RecommendationResponse"
      },
...
    }
  }
}